+++ packages/ttml-adapter/src/Parser/namespaces/tts/syntax/zIndex.ts
@@ -1,6 +1,6 @@
-import { integer } from "../structure/derivables/integer";
-import { keyword } from "../structure/derivables/keyword";
-import { oneOf } from "../../../structure/grammar";
+import { integer } from "../structure/derivables/integer.js";
+import { keyword } from "../structure/derivables/keyword.js";
+import { oneOf } from "../../../structure/grammar.js";
 
 /**
  * @syntax \<z-index>